// src/app/constants/enums/controls/form/form-validation-type.enum.ts
export enum ValidationType {
  REQUIRED = 'required',
  PASSWORD = 'password',
  REQUIRED_CHECKBOX = 'required-checkbox',
  DEFAULT_INPUT = 'default-input',
  POSITIVE = 'positive',
  POSITIVE_FROM_1 = 'positive-from-1',
  NEGATIVE = 'negative',
  MIN_LENGTH_1 = 'minlength-1',
  MIN_LENGTH_3 = 'minlength-3',
  MAX_LENGTH_100 = 'maxlength-100',
  MAX_LENGTH_30 = 'maxlength-30',
  MAX_LENGTH_20 = 'maxlength-20',
  MAX_LENGTH_400 = 'maxlength-400',
  MIN_1 = 'min-1',
  MAX_100 = 'max-100',
  MAX_20 = 'max-20',
  MAX_LENGTH_3 = 'maxlength-3',
  NO_FUTURE_DATE = 'no-future-date',
  NO_PAST_DATE = 'no-past-date',
  GREATER_THAN_DATE = 'greater-than-date',
  LESS_THAN_DATE = 'less-than-date',
  MAX_FILE_SIZE_2_MB = 'maxsize-2mb',
  MAX_FILE_SIZE_5_MB = 'maxsize-5mb',
  MAX_FILE_SIZE_10_MB = 'maxsize-10mb',
  MIN_FILE_1 = 'minfile-1',
  MIN_FILE_2 = 'minfile-2',
  MAX_FILE_1 = 'maxfile-1',
  MAX_FILE_5 = 'maxfile-5',
  MAX_FILE_10 = 'maxfile-10',
}